requirements.txt
streamlit
pillow
folium
streamlit-folium
psycopg2-binary
